Make MessageTranslator abstract
